+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Add [multi-region mode](https://docs.imgproxy.net/latest/serving_files_from_s3?id=multi-region-mode) to S3 integration.
 - Add the ability to [load environment variables](https://docs.imgproxy.net/latest/loading_environment_variables) from a file or a cloud secret.
 - Add `IMGPROXY_WORKERS` alias for the `IMGPROXY_CONCURRENCY` config.
+- (pro) Add [pages](https://docs.imgproxy.net/latest/generating_the_url?id=pages) processing option.
 
 ### Change
 - Don't report `The image request is cancelled` errors.